refactor: clean up files in matching-service
