+++ apps/matching-service/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"matching-service/handlers"
@@ -10,7 +9,6 @@ import (
 	"os"
 
 	"github.com/joho/godotenv"
-	"github.com/redis/go-redis/v9"
 )
 
 func main() {
@@ -19,33 +17,17 @@ func main() {
 	if err != nil {
 		log.Fatalf("err loading: %v", err)
 	}
-	port := os.Getenv("PORT")
-
-	// Retrieve redis url env variable and setup the redis client
-	redisAddr := os.Getenv("REDIS_URL")
-	client := redis.NewClient(&redis.Options{
-		Addr:     redisAddr,
-		Password: "", // no password set
-		DB:       0,  // use default DB
-	})
-
-	// Ping the redis server
-	_, err = client.Ping(context.Background()).Result()
-	if err != nil {
-		log.Fatalf("Could not connect to Redis: %v", err)
-	} else {
-		log.Println("Connected to Redis at the following address: " + redisAddr)
-	}
-
-	// Set redis client
-	processes.SetRedisClient(client)
+	
+	// Setup redis client
+	processes.SetupRedisClient()
 
 	// Run a goroutine that matches users
-
+	
 	// Routes
 	http.HandleFunc("/match", handlers.HandleWebSocketConnections)
-
+	
 	// Start the server
+	port := os.Getenv("PORT")
 	log.Println(fmt.Sprintf("Server starting on :%s", port))
 	err = http.ListenAndServe(fmt.Sprintf(":%s", port), nil)
 	if err != nil {
